Simple Profanity Filter Extended
Now I use Selenium WebDriver to do a particular action depending on if it found profanity or not. If no profanity was found, the test fails and we move on with our lives. If it found profanity, it continues to click the login button. I could easily replace that with Selenium flagging the post but I do not have a way to flag posts on the website.